// projeto/web/angular_p2/src/api/generated/models/eProductStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Pizzaria API
 * API para gerenciamento de pedidos e clientes.
 * OpenAPI spec version: 1.0
 */

export type EProductStatus = (typeof EProductStatus)[keyof typeof EProductStatus];

export const EProductStatus = {
  Active: 'Active',
  Inactive: 'Inactive',
} as const;