// projeto/web/angular_p2/src/api/generated/models/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Pizzaria API
 * API para gerenciamento de pedidos e clientes.
 * OpenAPI spec version: 1.0
 */
import type { Category } from './category';
import type { EProductStatus } from './eProductStatus';

export interface Product {
  name?: string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  price?: number | string;
  description?: string;
  /** @nullable */
  banner?: string | null;
  status?: EProductStatus;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  categoryId?: number | string;
  category?: Category;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  id?: number | string;
  createdAt?: string;
  updatedAt?: string;
}